+++ src/lib.rs
@@ -11,15 +11,16 @@ mod dto_factory;
 // Module containing service interfaces and implementations
 mod service;
 mod vo_factory;
+mod client;
 
 /// The main library module for the Lynn TCP Server.
 ///
 /// This module exports various components and utilities for building and running a TCP server.
 pub mod server {
-    // Exports the LynnConfigBuilder for configuring the server
-    pub use super::app::lynn_config_api::LynnConfigBuilder;
-    // Exports the LynnConfig for configuring the server
-    pub use super::app::lynn_config_api::LynnConfig;
+    // Exports the LynnServerConfigBuilder for configuring the server
+    pub use super::app::lynn_config_api::LynnServerConfigBuilder;
+    // Exports the LynnServerConfig for configuring the server
+    pub use super::app::lynn_config_api::LynnServerConfig;
     // Exports the LynnServer for running the server
     pub use super::app::LynnServer;
     // Exports the HandlerResult for handling the results of server operations
